+++ cosmovisor/cmd/cosmovisor/cmd/root.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"fmt"
 	"os"
-	"strings"
 
 	"github.com/rs/zerolog"
 	"github.com/spf13/cobra"
@@ -13,12 +12,11 @@ var logger *zerolog.Logger
 
 var rootCmd = &cobra.Command{
 	Use:   "cosmovisor",
-	Short: "Facilitate your cosmos-sdk upgrades with Cosmovisor",
-	Long:  "Cosmovisor is a small process manager for Cosmos SDK application binaries that monitors the governance module for incoming chain upgrade proposals.",
+	Short: "A process manager for Cosmos SDK application binaries.",
+	Long:  GetHelpText(),
 }
 
-// Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
+// Execute the CLI application.
 func Execute(log *zerolog.Logger) {
 	logger = log
 
@@ -26,14 +24,4 @@ func Execute(log *zerolog.Logger) {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-}
-
-// isOneOf returns true if the given arg equals one of the provided options (ignoring case).
-func isOneOf(arg string, options []string) bool {
-	for _, opt := range options {
-		if strings.EqualFold(arg, opt) {
-			return true
-		}
-	}
-	return false
 }