+++ cosmovisor/cmd/cosmovisor/main.go
@@ -5,8 +5,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/cosmovisor/cmd/cosmovisor/cmd"
 )
 
-var logger = cosmovisor.NewLogger()
-
 func main() {
-	cmd.Execute(logger)
+	cmd.Execute(cosmovisor.NewLogger())
 }